Accept different ways of specifying array type
This commit ensures that the creators for Array accept an interface{}
for the type argument. The type can be nil, the zero value of a go
slice, a dgo.ArrayType, or a dgo string that parses to a dgo.ArrayType.
A new ArrayWithCapacity is also added.

The change makes the Array creators more consistent with the Map
creators.