Ensure that Function.Call doesn't translate a return of nil into Nil.
In general, Nil represents a valid value in dgo (like null in javascript)
whereas nil represents nothing at all (like undefined in javascript). A
Nil can be stored in a map. A nil however cannot. When a map returns a
nil, that means that there was no association for the given key.

A Function must be allowed to return a nil to represent a similar "not
found" condition. This commit ensures that such a nil is propagated
correctly without being transformed into a Nil.