Remove use of json.Unmarshaler and json.Marshaler interfaces.
This commit removes the use of the `json.Marshaler` and
`json.Unmarshaler` since they introduced behaviors that are inconsistent
with the new streamer and moves the current `vf.MarshalJSON()` and
`vf.UnmarshalJSON()` to the streamer package.

Users are encouraged to use the streamer package when performing JSON
operations on dgo values. The alternative is to use the bidirectional
conversion capabilites `vf.Value()` and `vf.FromValue() and then use
the runtime json package for the rest. That works fine in all cases
where the data can be transformed to JSON without type loss but it does
not work for Timestamp, Sensitive, etc.