Add ability to add named types with constructors and init arguments.
This commit adds a new `NamedType` to the type system. A named type is
a type that is represented by a name and made known to the system by
associating a name with a set of functions that can determine
assignability, init arguments and creation.

Implementation of parameterized named types is not included in this
commit.