Add new Arguments type
The Arguments type is used when there's a need to send more than one
argument to a factory method. It's essentially the same as an Array
but gives the receiver a chance to differentiate between one argument
that is an Array and multiple arguments where the first can be an
Array.