Improve stringer to handle aliases
The stringer.TypeString method currently expands all types as much as
possible and will emit the string "<recursive self reference to ...>"
whenever it encounters a self referencing alias.

This commit changes the TypeString method so that it instead uses the
`AliasMap` obtained from the function `DefaultAliases` to resolve types
into their respective alias names, thereby enabling types like `data`
to be printed as "data" rather than with the aforementioned recursive
message.

Since DefaultAliases is also used as the default AliasMap when parsing
or deserializing, the aliases added when doing that will be recognized
by the TypePrinter.