Remove ability to have a type embedded in Map and Array
Prior to this commit, a Map or an Array could have an embedded type
that would override the default type (the type backed by the collection
itself). The reason for this was to allow "typed collections", i.e.
collections that wouldn't accept modifications beyond what its type
would allow.

The approach with typed collections led to ambiguities and overly
complex code. It was also never used in any of dgo's applications since
the most common place to check types when assigning to variables or
passing parameters. The ability to have embedded types in collections
is therefore removed by this commit.