Let serializer recognize and propagate references to collector
The serializer would propagate a map in the form {"__ref":4} to the
collector as any other map. This was wrong. The serializer must call
the collectors `AddRef` method instead of its `AddMap` when such a map
is encountered. Not doing so means loss of ref semantics. This commit
ensures that `AddRef` is called correctly.

Closes lyraproj/hiera#74