Add JSON decoded that retains order of object entries in Map
In order for `{ __ref: <n> }` constructs to work properly, the sequence
of values in a JSON stream must be retained throughout. The standard
Go JSON decoder uses a `map[string]interface{}` internally and will
therefore map entries in random order. This commit adds a decoder that
creates a dgo.Map directly from the stream and thus, retains the order
in which the entries are added.