Remove exact types (make values be their own type)
This commit removes almost all exact types in favor of instead letting
values implement the type interface and hence be their own type
representation. One exact type remains and that's the one used for
representing named types (native types for which a dgo.Type is created).

This commit also changes the organisation from 'lyraproj' to 'tada' and
removes the Pcore specific type parser. The changes are not intended to
be backward compatible. This commit should be considered the start of a
new fork of the lyraproj/dgo project which is not intended to be used
with Puppet's go version of hiera.