Add BigInt and BigFloat data types
This commit adds the data types BigInt and BigFloat to the set of
types managed by Dgo. The parser introduces a new keyword "big"
that can be uses as a prefix for a float or an integer literal in
order to enforce parsing into a BigFloat or a BigInt.

When parsing integer literals, the parser will automatically
select the correct type based on the magnitude of the parsed
integer. This does however not happen with BigFloat since the
magnitude isn't the most important thing for a BigInt. It's
the desired precision and the precision has no correlation to
a decimal string representation.